refactor: decompose errors module
