+++ task_01_ciphering_cli_tool/errors/pipelineErrors.js
@@ -1,0 +1,12 @@
+const { CustomErrors } = require('./customErrors');
+
+class PipelineError extends CustomErrors {
+    constructor(message) {
+        super(message);
+        this.name = 'PipelineError';
+    };
+};
+
+module.exports = {
+    PipelineError
+};
