+++ task_01_ciphering_cli_tool/index.js
@@ -1,6 +1,6 @@
 const { cliValidation } = require('./validation/index');
 const { createPipeline } = require('./streams/createPipeline');
-const { errorHandler } = require('./errors/customErrors');
+const { errorHandler } = require('./errors/index');
 
 const cliParameters = process.argv.slice(2);
