+++ task_01_ciphering_cli_tool/streams/createPipeline.js
@@ -2,7 +2,7 @@ const { pipeline } = require('stream');
 const { createInputStream } = require('./inputStream');
 const { createOutputStream } = require('./outputStream');
 const { createTransformStreamCollection } = require('./transformStream');
-const { errorHandler, PipelineError } = require('../errors/customErrors');
+const { errorHandler, PipelineError } = require('../errors/index');
 
 const createPipeline = (cliParameters) => {
     const inputStream = createInputStream(cliParameters);
